workaround: gn_to_cmakelist.py trim cmake names to a maximum length of 215 characters.
For every CMake target a corresponding file is created on the file system.
Typical file systems have a maxiumum file name length of 255 characters. This is easily exceeded
with the chip project configuration.
This employs a workaround by trimming the target_name, though not checking currently for any conflicts
introduced. So consider this a workaround only applicable to this project.